Don't destroy existing location.
This is a bodge - if we have location data for a user, and regardless of
what the LDAP lookup finds, just keep it rather than overwriting with
nothing. If we used the LDAP lookup for locations I'd modify this to
only restore the saved location if the set one was 0 - ie. the lookup
didn't find anything.